do not ignore OOB on toptag=0
this was a temporary fix when globals weren't properly being ignored yet,
and stack objects didnt have a proper in-pointer tag yet.

Since then, both of those issues have been fixed, and ignoring the toptag 0
causes false negatives for stack objects that happen to be allocated at
the start of an arena.

Ideally, we would still avoid allocating stack objects at the edges of
arenas. But for now, it doesnt seem to cause problems on SPEC