add optional 'ignore errors' config, default enable it on AMD UAI
Ignore errors by "forcing" the NON_NOOB_MIN_RADIX value to be looked up from pointers in instrumentation,
regardless of the actual radix value. We also force a data-dependency on some opaque (but known to us as 0) bits
to avoid the compiler optimizing based on the known radix value.

This should give at least as much operations/overhead as the actual instrumentation. But it allows us
to leave the top bits blank and still benchmark accurately.

Proper UAI probably really needs native kernel support.
We tried working around it using lazypoline, by masking out the top bits on pointer arguments
But we ran into bugs in lazypoline. Another syscall interposer might work, but we ran into a second issue:
SIGBUS signals when accessing some NOOB memory

We suspect that, when demand-paging through NOOB-managed memory, page faults on tagged pointers
make the kernel go crazy. We _could_ perhaps solve this with a userfault handler, but ignoring the errors is easier.
And may allow us to make the ASPLOS deadline.